refactor(board-rdb-adapter): use base-package as `nettee.board.driven.rdb`
