+++ services/board/driven/rdb/src/main/java/nettee/board/driven/rdb/BoardJpaRepository.java
@@ -1,6 +1,6 @@
-package nettee.board;
+package nettee.board.driven.rdb;
 
-import nettee.board.entity.BoardEntity;
+import nettee.board.driven.rdb.entity.BoardEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BoardJpaRepository extends JpaRepository<BoardEntity, Long> { }
